remain the old handler to keep forward compatibility
