--- libcpu/arm/at91sam926x/interrupt.h
@@ -1,11 +1,0 @@
-#ifndef __INTERRUPT_H__
-#define __INTERRUPT_H__
-
-struct rt_irq_desc {
-	char irq_name[RT_NAME_MAX];
-	rt_isr_handler_t isr_handle;
-	void *param;
-	rt_uint32_t interrupt_cnt;
-};
-
-#endif
